Fixed ServerContextImpl::isClientEcdsaCapable() for TLS 1.3 clients
The client_version local variable is intended to be the value of the legacy version field from
the client hello message, which we can obtain with SSL_client_hello_get0_legacy_version().

Previously we were incorrectly taking the value from ssl->version which was a naive
modification of the original upstream code which used client_hello->version.

Signed-off-by: Ted Poole <tpoole@redhat.com>